requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
python-decouple
fastapi-mail
passlib[bcrypt]
